python2/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
jieba==0.39
pyparsing==2.0.1
PyYAML==3.12
requests==2.18.4
scipy==0.13.0b1
urllib3==1.22
Whoosh==2.7.4
xattr==0.6.4
